fix: build error

+++ internal/link/device_unix.go
@@ -11,7 +11,7 @@ import (
 // This is used to check if two paths are on the same filesystem
 func getDeviceID(info os.FileInfo) uint64 {
 	if stat, ok := info.Sys().(*syscall.Stat_t); ok {
-		return stat.Dev
+		return uint64(stat.Dev)
 	}
 	return 0
 }
